fix: daysToConversion schema IntegerType → DoubleType (contains decimals like 1.88)

+++ lakehouse/models/bronze/schema/shopify_orders_schema.py
@@ -132,7 +132,7 @@ SHOPIFY_ORDERS_SCHEMA = StructType([
         # Customer journey — FIX ❌: lastVisit was StringType, confirmed struct in data
         StructField("customerJourneySummary", StructType([
             StructField("customerOrderIndex", IntegerType(), True),
-            StructField("daysToConversion",   IntegerType(), True),
+            StructField("daysToConversion",   DoubleType(),  True),
             StructField("momentsCount", StructType([
                 StructField("count",     IntegerType(), True),
                 StructField("precision", StringType(),  True),
